requirements.txt
python-dotenv
werkzeug==2.1.2
Flask~=2.1.2
flask-restx==0.5.1
uWSGI
uwsgi-tools
click<8.1.0
pysqlite3
redis==4.3.4
Flask-Caching==2.0.1
Flask-Limiter
requests